backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pdfminer.six>=20221105
python-docx>=1.1.0
pillow>=10.0.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
pydantic>=2.0.0
tqdm>=4.65.0
types-tqdm>=4.65.0
pandas>=2.0.0
openpyxl>=3.1.0
PyMuPDF>=1.23.0
playwright>=1.40.0
beautifulsoup4>=4.14.0
